pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlbydoing"
version = "0.1.0"
description = "Deep learning by doing: hands-on PyTorch experiments"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "torch-geometric>=2.5",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
